storage: refactor storage_cfg()
It was almost entirely done in a single function, was too
difficult to extend in next commits. This patch splits the config
code into several functions doing a certain step of the
configuration.

The patch also introduces a config context. It is an object where
one time per cfg call are calculated and saved all the most useful
config details such as the current replica and replicaset UUID,
whether the instance was a master before and is now, config of
this particular instance and replicaset, and more.

The context is introduced because these attributes are going to be
used more widely across several functions, and passing them as
arguments everywhere would be just unhandy. As well as calculating
them again in each function.

For example, it allows storage_cfg_master_prepare() and
storage_cfg_master_commit() just use the already provided
was_master_in_cfg and is_master_in_cfg and concentrate on the
switch logic instead of figuring out how did the state change.

Part of #429

NO_DOC=refactoring
NO_TEST=refactoring