storage: introduce recovery_bucket_stat()
The new function is used by bucket recovery instead of
bucket_stat(). The problem of the latter is that it allows to be
invoked on a non-master instance which is not ok for recovery - it
needs to get the most up to date state of the bucket.

On the other hand it wasn't possible to just ban bucket_stat() on
non-master instances, because it is a public function and it was
never documented to work exclusively on master instances.

Part of #430

NO_DOC=internal change