storage: move master and rebalancer role commit
It was done after schema upgrade, but now is done before. It makes
sense because master commit should enable read-write mode which is
needed for the schema changes.

It worked so far, because in manual master config the read-only
cfg update is done at the box.cfg stage. I.e. before the schema
update.

But soon the master role will be allowed to turn on and off
automatically. And then the old order wouldn't work anymore.

Part of #429

NO_DOC=refactoring
NO_TEST=refactoring