storage: add timeouts to all master_call()
Many of them used to use the default timeout
(replica.net_timeout). It is working good enough, but isn't really
suitable for storages - this auto-tuning timeout thing was
designed for routers mostly.

Besides, soon in master_call() it will be needed to know the
timeout before doing the call itself. Getting the net_timeout
out of the master object would work for now, but soon this won't
be possible when auto-master will be enabled with discovery.

That all means either trying schemas like the router does for the
same job, or simply always pass an explicit timeout to all
master calls. The patch follows the latter way.

The need to know the timeout will appear for doing retries inside
master_call() when the master is not known right now, but chances
are high is going to be discovered in a few milliseconds. For the
waiting a deadline is needed, and for that a timeout is necessary.

Part of #430

NO_DOC=internal change
NO_TEST=covered